+++ DependencyInjection/Compiler/RequestAwarePass.php
@@ -19,8 +19,8 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 
 /**
- * A compiler pass to find all cmf_request_aware services and add them to the
- * RequestAwareListener
+ * BC: A compiler pass to find all cmf_request_aware services and adjust the
+ * service definition.
  *
  * @author David Buchmann <mail@davidbu.ch>
  */
@@ -31,43 +31,12 @@ class RequestAwarePass implements CompilerPassInterface
      */
     public function process(ContainerBuilder $container)
     {
-        if (version_compare(Kernel::VERSION, '2.3', '<')) {
-            $this->configureSynchronizer($container);
-        } else {
-            $this->makeSynchronized($container);
-        }
-    }
-
-    /**
-     * Configure the request synchronizer for symfony 2.2
-     *
-     * @param ContainerBuilder $container
-     */
-    private function configureSynchronizer(ContainerBuilder $container)
-    {
-        if (!$container->hasDefinition('cmf_core.listener.request_aware')) {
-            return;
-        }
-
-        $listener = $container->getDefinition('cmf_core.listener.request_aware');
         $services = $container->findTaggedServiceIds('cmf_request_aware');
         foreach ($services as $id => $attributes) {
-            $listener->addMethodCall('addService', array(new Reference($id)));
-        }
-    }
+            trigger_error("Service $id is using the deprecated tag cmf_request_aware");
 
-    /**
-     * Make the tagged services synchronized for symfony 2.3 and later.
-     *
-     * @param ContainerBuilder $container
-     */
-    private function makeSynchronized(ContainerBuilder $container)
-    {
-        $services = $container->findTaggedServiceIds('cmf_request_aware');
-        foreach ($services as $id => $attributes) {
             $definition = $container->getDefinition($id);
             $definition
-                ->setSynchronized(true)
                 ->addMethodCall('setRequest', array(
                     new Reference('request', ContainerInterface::NULL_ON_INVALID_REFERENCE, false)
                 ))